+++ CHANGELOG.rst
@@ -25,6 +25,8 @@ Backwards incompatible changes
 Added
 ^^^^^
 - :meth:`tenpy.networks.mps.MPS.canonical_form_infinite`.
+- :meth:`tenpy.netwoks.mps.MPS.expectation_value_term`, :meth:`tenpy.netwoks.mps.MPS.expectation_value_terms_sum` and
+  :meth:`tenpy.netwoks.mps.MPS.expectation_value_multi_site` for expectation values of terms.
 - :meth:`tenpy.linalg.np_conserved.Array.extend` and :meth:`tenpy.linalg.charges.LegCharge.extend`,
   allowing to extend an Array with zeros.
 - DMRG parameter ``'orthogonal_to'`` allows to calculate excited states for finite systems.